v2.2.0 (#26)
* Add test for Case-Insensitive operator at the beginning

* update benchmark

* change page type to int - issue #23

* fix bug #27

* add Processing null or default feature #25

* add Ef Core Comparison operations Test (in-memory provider) #27

* Add EntityFramework Compatibility Layer #27

* update to v2.2.0

+++ src/Gridify/IGridifyPagination.cs
@@ -2,7 +2,7 @@ namespace Gridify
 {
    public interface IGridifyPagination
    {
-      short Page { get; set; }
+      int Page { get; set; }
       int PageSize { get; set; }   
    }
 }
